unified implementation of http client package division
