+++ common/src/main/java/com/alibaba/nacos/common/http/client/handler/StringResponseHandler.java
@@ -14,28 +14,26 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.common.http.client;
+package com.alibaba.nacos.common.http.client.handler;
 
 import com.alibaba.nacos.common.http.HttpRestResult;
+import com.alibaba.nacos.common.http.client.response.HttpClientResponse;
 import com.alibaba.nacos.common.http.param.Header;
-import com.alibaba.nacos.common.utils.JacksonUtils;
+import com.alibaba.nacos.common.utils.IoUtils;
 
-import java.io.InputStream;
 import java.lang.reflect.Type;
 
 /**
- * bean response handler,
- * Mainly converter response type as bean type.
+ * string response handler, Mainly converter response type as string type.
  *
  * @author mai.jh
  */
-public class BeanResponseHandler<T> extends AbstractResponseHandler<T> {
+public class StringResponseHandler extends AbstractResponseHandler<String> {
     
     @Override
-    public HttpRestResult<T> convertResult(HttpClientResponse response, Type responseType) throws Exception {
+    public HttpRestResult<String> convertResult(HttpClientResponse response, Type responseType) throws Exception {
         final Header headers = response.getHeaders();
-        InputStream body = response.getBody();
-        T extractBody = JacksonUtils.toObj(body, responseType);
-        return new HttpRestResult<T>(headers, response.getStatusCode(), extractBody, null);
+        String extractBody = IoUtils.toString(response.getBody(), headers.getCharset());
+        return new HttpRestResult<String>(headers, response.getStatusCode(), extractBody, null);
     }
 }
