+++ common/src/main/java/com/alibaba/nacos/common/http/client/request/HttpClientRequest.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.common.http.client;
+package com.alibaba.nacos.common.http.client.request;
 
+import com.alibaba.nacos.common.http.client.response.HttpClientResponse;
 import com.alibaba.nacos.common.model.RequestHttpEntity;
 
 import java.io.Closeable;
